OBE result API for Faculty & UI Completed
Implemented OBE result API for faculty to fetch CLO-based student attainment.- Optimized API response structure for efficient data retrieval.- Fixed assessment contribution calculations and ensured correct mappings.- Designed Vue.js-based UI with Select Section, CLO Table, and Student Table components.- Improved table responsiveness with Bootstrap.- Extracted proper assessment types dynamically from API.- Fixed data binding issues in Student Table component.- Structured Vue files with best practices for modularity.- Finalized API integration with the UI for seamless data flow. Next Steps:- Fix remaining issue in Student Table (marks display issue).- Final review and optimization for production readiness.

+++ assessments/models/assessment_breakdown.py
@@ -6,12 +6,12 @@ class AssessmentBreakdown(models.Model):
     section = models.OneToOneField(Section, on_delete=models.CASCADE)  # Ensures only one breakdown per section
 
     # Individual fields for each assessment type
-    assignment_weightage = models.PositiveIntegerField(default=0)
-    quiz_weightage = models.PositiveIntegerField(default=0)
-    lab_weightage = models.PositiveIntegerField(default=0)
-    mid_weightage = models.PositiveIntegerField(default=0)
-    final_weightage = models.PositiveIntegerField(default=0)
-    project_weightage = models.PositiveIntegerField(default=0)
+    assignment_weightage = models.FloatField(default=0)
+    quiz_weightage = models.FloatField(default=0)
+    lab_weightage = models.FloatField(default=0)
+    mid_weightage = models.FloatField(default=0)
+    final_weightage = models.FloatField(default=0)
+    project_weightage = models.FloatField(default=0)
 
     def clean(self):
         # Validate that the total weightage equals 100
